+++ apps/omsorgsdager-kalkulator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/omsorgsdager-kalkulator
 
+## 3.6.6
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+
+## 3.6.5
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+
 ## 3.6.4
 
 ### Patch Changes