+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-soknad-ds
 
+## 10.15.7
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+
+## 10.15.6
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+
 ## 10.15.5
 
 ### Patch Changes