+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/sif-common-ui
 
+## 0.7.8
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.8.7
+    -   @navikt/sif-common-core-ds@9.17.8
+    -   @navikt/sif-common-utils@3.46.2
+
+## 0.7.7
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.8.6
+    -   @navikt/sif-common-core-ds@9.17.7
+    -   @navikt/sif-common-utils@3.46.1
+
 ## 0.7.6
 
 ### Patch Changes